Cleanup old tested gradle runner versions
The GradleRunner tests had a number of annotations marking when certain features were introduced
in TestKit. Many of these versions corresponded to Gradle versions before 4.0 -- the min supported
Gradle version in Gradle 9.0

We remove the older annotations, leaving only two remaining.

In addition, we fix the deprecation warning that is emitted when the TestKit is executed on an
unsupported target Gradle version. Previously, we used the DeprecationLogger. However, the
logger is only properly initialized in the Gradle daemon, meaning the warnings were
never emitted to the console.

We now use an SLF4J logger to emit the warning, and add tests to verify this warning
is emitted via the logging infrastructure.