Ceanup NodeState
After #33925, we follow up with some refactorings. Most importantly, we clean up
visitOutgoingDependencies and excludesSameDependenciesAsPreviousTraversal, which were two
very coupled methods. excludesSameDependenciesAsPreviousTraversal performed side-effects
which anticipated what visitOutgoingDependencies would do with its return value.

This sort of coupling is confusing to follow and maintain. We inline and refactor some logic
to make the logic easier to follow and maintain.